Fixed minimum version in filenames and fixed muzzle check
