+++ instrumentation/spring/spring-rmi-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/springrmi/server/ServerInstrumentationModule.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.springrmi.client;
+package io.opentelemetry.javaagent.instrumentation.springrmi.server;
 
 import static java.util.Collections.singletonList;
 
@@ -13,14 +13,14 @@ import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import java.util.List;
 
 @AutoService(InstrumentationModule.class)
-public class ClientInstrumentationModule extends InstrumentationModule {
+public class ServerInstrumentationModule extends InstrumentationModule {
 
-  public ClientInstrumentationModule() {
-    super("spring-rmi", "spring-rmi-client");
+  public ServerInstrumentationModule() {
+    super("spring-rmi", "spring-rmi-4.0", "spring-rmi-server");
   }
 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new ClientInstrumentation());
+    return singletonList(new ServerInstrumentation());
   }
 }
