+++ settings.gradle.kts
@@ -367,7 +367,7 @@ include(":instrumentation:spring:spring-integration-4.1:library")
 include(":instrumentation:spring:spring-integration-4.1:testing")
 include(":instrumentation:spring:spring-kafka-2.7:javaagent")
 include(":instrumentation:spring:spring-rabbit-1.0:javaagent")
-include(":instrumentation:spring:spring-rmi:javaagent")
+include(":instrumentation:spring:spring-rmi-4.0:javaagent")
 include(":instrumentation:spring:spring-scheduling-3.1:javaagent")
 include(":instrumentation:spring:spring-web-3.1:javaagent")
 include(":instrumentation:spring:spring-web-3.1:library")
